Support nesting of Bundler block DSL
