+++ lib/appraisal/group.rb
@@ -1,50 +1,23 @@
-require 'appraisal/dependency_list'
+require 'appraisal/bundler_dsl'
 require 'appraisal/utils'
 
 module Appraisal
-  class Group
+  class Group < BundlerDSL
     def initialize(group_names)
-      @dependencies = DependencyList.new
+      super()
       @group_names = group_names
-      @gemspec = nil
-    end
-
-    def run(&block)
-      instance_exec(&block)
-    end
-
-    def gem(name, *requirements)
-      @dependencies.add(name, requirements)
-    end
-
-    def gemspec(options = {})
-      @gemspec = Gemspec.new(options)
     end
 
     def to_s
-      formatted_output dependencies_list
+      formatted_output indent(super)
     end
 
     # :nodoc:
     def for_dup
-      formatted_output dependencies_list_for_dup
+      formatted_output indent(super)
     end
 
     private
-
-    def dependencies_list
-      Utils.join_parts([
-        dependencies_entry,
-        gemspec_entry
-      ]).gsub(/^/, "  ")
-    end
-
-    def dependencies_list_for_dup
-      Utils.join_parts([
-        dependencies_entry,
-        gemspec_entry_for_dup
-      ]).gsub(/^/, "  ")
-    end
 
     def formatted_output(output_dependencies)
       <<-OUTPUT.strip
@@ -52,20 +25,6 @@ group #{Utils.format_arguments(@group_names)} do
 #{output_dependencies}
 end
       OUTPUT
-    end
-
-    def dependencies_entry
-      @dependencies.to_s
-    end
-
-    def gemspec_entry
-      @gemspec.to_s
-    end
-
-    def gemspec_entry_for_dup
-      if @gemspec
-        @gemspec.for_dup
-      end
     end
   end
 end
